# docs/swagger.yaml
basePath: /v1/user-services
definitions:
  apperror.Code:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - NOT_ACCEPTABLE
    - REQUEST_TIMEOUT
    - CONFLICT
    - GONE
    - PRECONDITION_FAILED
    - PAYLOAD_TOO_LARGE
    - UNSUPPORTED_MEDIA_TYPE
    - VALIDATION_ERROR
    - UNPROCESSABLE_ENTITY
    - LOCKED
    - FAILED_DEPENDENCY
    - PRECONDITION_REQUIRED
    - TOO_MANY_REQUESTS
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    - BAD_GATEWAY
    - SERVICE_UNAVAILABLE
    - TIMEOUT
    - CACHE_MISS
    - CACHE_ERROR
    - MESSAGE_BUS_ERROR
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeNotAcceptable
    - CodeRequestTimeout
    - CodeConflict
    - CodeGone
    - CodePrecondFailed
    - CodePayloadTooLarge
    - CodeUnsupportedMedia
    - CodeValidation
    - CodeUnprocessable
    - CodeLocked
    - CodeFailedDependency
    - CodePrecondRequired
    - CodeTooManyRequest
    - CodeInternal
    - CodeNotImplemented
    - CodeBadGateway
    - CodeServiceUnavail
    - CodeTimeout
    - CodeCacheMiss
    - CodeCacheError
    - CodeMessageBus
...
  github_com_modami_user-service_internal_dto.CreateReviewRequest:
    properties:
      comment:
        maxLength: 1000
        type: string
      is_anonymous:
        type: boolean
      order_id:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - order_id
    - rating
    type: object
  github_com_modami_user-service_internal_dto.FollowListResponse:
    properties:
      cursor:
        type: string
      users:
        items:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.FollowUserItem'
        type: array
    type: object
  github_com_modami_user-service_internal_dto.FollowStatusResponse:
    properties:
      is_following:
        type: boolean
    type: object
  github_com_modami_user-service_internal_dto.FollowUserItem:
    properties:
      avatar_url:
        type: string
      followed_at:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
...
  github_com_modami_user-service_internal_dto.RatingSummaryResponse:
    properties:
      avg_rating:
        type: number
      count_1:
        type: integer
      count_2:
        type: integer
      count_3:
        type: integer
      count_4:
        type: integer
      count_5:
        type: integer
      total_reviews:
        type: integer
      user_id:
        type: string
    type: object
...
  github_com_modami_user-service_internal_dto.ReviewListResponse:
    properties:
      cursor:
        type: string
      reviews:
        items:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.ReviewResponse'
        type: array
    type: object
  github_com_modami_user-service_internal_dto.ReviewResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_anonymous:
        type: boolean
      order_id:
        type: string
      rating:
        type: integer
      reviewee_id:
        type: string
      reviewer_id:
        type: string
      role:
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        $ref: '#/definitions/apperror.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
    type: object
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.Pagination'
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
  response.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8086
info:
  contact: {}
  description: User service for the Modami marketplace platform.
  title: Modami User Service API
  version: "1.0"
paths:
...
  /users/{id}/follow:
    delete:
      description: Unfollow a user by their ID
      parameters:
      - description: User ID to unfollow (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - Follows
    post:
      description: Follow another user by their ID
      parameters:
      - description: User ID to follow (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - Follows
  /users/{id}/follow/status:
    get:
      description: Check if the authenticated user follows another user
      parameters:
      - description: User ID to check (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.FollowStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Check follow status
      tags:
      - Follows
  /users/{id}/followers:
    get:
      description: Returns a paginated list of a user's followers
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.FollowListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get followers
      tags:
      - Follows
  /users/{id}/following:
    get:
      description: Returns a paginated list of users that a user follows
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.FollowListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get following
      tags:
      - Follows
  /users/{id}/reviews:
    get:
      description: Returns a paginated list of reviews for a user
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.ReviewListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: List reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Create a review for a user (as a buyer)
      parameters:
      - description: Reviewee user ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Review details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - Reviews
  /users/{id}/reviews/summary:
    get:
      description: Returns rating statistics for a user
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.RatingSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get rating summary
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"